+++ src/imgui.lua
@@ -1,6 +1,11 @@
 ---@meta _
 ---@diagnostic disable
 
+local previousConfig = {
+    dress = nil,
+    random_each_run = nil,
+}
+
 rom.gui.add_imgui(function()
     if rom.ImGui.Begin("Dress Selector") then
         drawMenu()
@@ -18,16 +23,27 @@ end)
 function drawMenu()
     rom.ImGui.Text("Select Dress")
     if rom.ImGui.BeginCombo("###dress", config.dress) then
-        for _, dressPair in ipairs(mod.DressData) do
-            local  dressName = dressPair[1]
-            local  dressValue = dressPair[2]
+        for _, dressName in ipairs(mod.DressDisplayOrder) do
             if rom.ImGui.Selectable(dressName, (dressName == config.dress)) then
-                config.dress = dressName
-                mod.dressvalue = dressValue
-                mod.UpdateSkin(mod.dressvalue)
+                if dressName ~= previousConfig then
+                    local  dressGrannyTexture = mod.GetDressGrannyTexture(dressName)
+                    config.dress = dressName
+                    previousConfig.dress = dressName
+                    config.random_each_run = false
+                    game.SetupCostume()
+                end
+                rom.ImGui.SetItemDefaultFocus()
             end
-            rom.ImGui.SetItemDefaultFocus()
         end
         rom.ImGui.EndCombo()
+    end
+    
+    rom.ImGui.Separator()
+
+    local value, checked = rom.ImGui.Checkbox("Random Dress Each Run", config.random_each_run)
+    if checked and value ~= previousConfig.random_each_run then
+        config.random_each_run = value
+        previousConfig.random_each_run = value
+        game.SetupCostume()
     end
 end