+++ src/main.lua
@@ -1,7 +1,7 @@
 ---@meta _
 -- grabbing our dependencies,
 -- these funky (---@) comments are just there
---	 to help VS Code find the definitions of things
+--     to help VS Code find the definitions of things
 
 ---@diagnostic disable-next-line: undefined-global
 local mods = rom.mods
@@ -9,7 +9,7 @@ local mods = rom.mods
 ---@module 'LuaENVY-ENVY-auto'
 mods['LuaENVY-ENVY'].auto()
 -- ^ this gives us `public` and `import`, among others
---	and makes all globals we define private to this plugin.
+--    and makes all globals we define private to this plugin.
 ---@diagnostic disable: lowercase-global
 
 ---@diagnostic disable-next-line: undefined-global
@@ -39,22 +39,37 @@ config = chalk.auto 'config.lua'
 public.config = config -- so other mods can access our config
 
 local function on_ready()
-	-- what to do when we are ready, but not re-do on reload.
-	if config.enabled == false then return end
-	mod = modutil.mod.Mod.Register(_PLUGIN.guid)
-	mod.dressvalue = ""
-	import 'ponydata.lua'
-	import 'ponylogic.lua'
-	import 'ready.lua'
+    -- what to do when we are ready, but not re-do on reload.
+    if config.enabled == false then return end
+    mod = modutil.mod.Mod.Register(_PLUGIN.guid)
+
+    import 'ponydata.lua'
+    import 'ponylogic.lua'
+    mod.PopulatePonyMenuData()
+
+    import 'data.lua'
+    import 'setupdata.lua'
+    import 'sjson.lua'
+    import 'ready.lua'
 end
 
 local function on_reload()
-	-- what to do when we are ready, but also again on every reload.
-	-- only do things that are safe to run over and over.
-	if config.enabled == false then return end
+    -- what to do when we are ready, but also again on every reload.
+    -- only do things that are safe to run over and over.
+    if config.enabled == false then return end
 
-	import 'reload.lua'
-	import 'imgui.lua'
+    import 'reload.lua'
+    import 'imgui.lua'
+
+    if config.debug_reload == false then return end
+
+    import 'ponydata.lua'
+    import 'ponylogic.lua'
+    import 'data.lua'
+    import 'setupdata.lua'
+
+    ModUtil.Table.Merge(ScreenData,mod.DressScreenData)
+    
 end
 
 -- this allows us to limit certain functions to not be reloaded.
@@ -62,5 +77,5 @@ local loader = reload.auto_single()
 
 -- this runs only when modutil and the game's lua is ready
 modutil.once_loaded.game(function()
-	loader.load(on_ready, on_reload)
+    loader.load(on_ready, on_reload)
 end)