+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+- Add option to favorite dress for randomization.
+- Add a preview window for Mel.
+
+## [1.4.0] - 2025-12-01
+
+- Add upgrade menu portrait
+- Update dialogue portraits for Lavender, Moonlight and Fuchsia
+
+## [1.3.0] - 2025-11-29
+
+- Add option to randomize dress at run start
+
 ## [1.2.1] - 2025-11-27
 
 - Add Murderrrrr skin by burn
@@ -50,7 +62,9 @@
 
 - First version of the mod!
 
-[unreleased]: https://github.com/adi1998/MelSkin/compare/1.2.1...HEAD
+[unreleased]: https://github.com/adi1998/MelSkin/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/adi1998/MelSkin/compare/1.3.0...1.4.0
+[1.3.0]: https://github.com/adi1998/MelSkin/compare/1.2.1...1.3.0
 [1.2.1]: https://github.com/adi1998/MelSkin/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/adi1998/MelSkin/compare/1.1.2...1.2.0
 [1.1.2]: https://github.com/adi1998/MelSkin/compare/1.1.1...1.1.2